sdm660-common: Fix dac* denials for persist forced restorecon
Unfortunately, Moto sdm660-common devices have too many permissions which would cause the command
to fail if we added them all at once, so we must break up the commands depending on the folder/file
in the root persist directory.

We must also fix group permissions for a few persist folders as there are
some that have permissions that only exist on stock and change the command used for restorecon
to ensure it is ran forcibly while still being recursive as restorecon_recursive isn't forced.

Plus fix some sepolicy denials that popped up

Change-Id: Ide9f37947a6e4ed6712b1d17265954f6eb260da3